requirements.txt
Flask==2.2.5
Werkzeug==2.2.3
Flask-Login==0.6.2
Flask-SQLAlchemy==3.0.3
Flask-WTF==1.1.1
WTForms==3.0.1
Flask-Migrate==4.0.4
email-validator==1.3.1
python-dotenv==1.0.0
flask-bcrypt==1.0.1
Flask-Session==0.5.0
reportlab==4.0.4
openpyxl==3.1.2
pandas==2.0.3

